requirements.txt
# Core ADK and Agent Dependencies
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.2
vertexai>=1.46.0

# CrewAI for Multi-Agent Orchestration
crewai>=0.30.0
crewai-tools>=0.2.0

# Google Cloud Services
google-cloud-vision>=3.4.0
google-cloud-translate>=3.12.0
google-cloud-firestore>=2.11.0
google-cloud-storage>=2.10.0
google-cloud-retail>=1.16.0
google-cloud-discoveryengine>=0.11.0

# LangChain and Vector Store
langchain>=0.1.0
langchain-google-vertexai>=0.0.6
langchain-google-firestore>=0.0.1
chromadb>=0.4.22

# Image Processing
Pillow>=10.1.0
opencv-python>=4.8.1

# API Framework
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Data Processing
pandas>=2.1.3
numpy>=1.24.3
scikit-learn>=1.3.2

# Utils
python-dotenv>=1.0.0
httpx>=0.25.0
redis>=5.0.1
celery>=5.3.4

# Development & Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
pylint>=3.0.2

# Monitoring
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
prometheus-client>=0.19.0